+++ front-end/app/bootstrap/bootstrap.component.ts
@@ -1,0 +1,23 @@
+import { Component, OnInit, AfterContentInit } from "@angular/core";
+
+@Component({
+  selector: "bootstrap",
+  template: ""
+})
+export class BootstrapComponent implements OnInit, AfterContentInit {
+
+  constructor() {}
+
+  ngOnInit() {
+    var f = this;
+
+    //
+  }
+
+  ngAfterContentInit() {
+    var f = this;
+    
+    //
+  }
+  
+}