+++ front-end/app/routes/home/home.component.ts
@@ -1,10 +1,10 @@
-import { Component, AfterContentInit } from "@angular/core";
+import { Component } from "@angular/core";
 import { Title } from "@angular/platform-browser";
 
 @Component({
   templateUrl: "/app/routes/home/home.html"
 })
-export class HomePageComponent implements AfterContentInit {
+export class HomePageComponent {
 
   constructor(
     public title:Title
@@ -12,12 +12,6 @@ export class HomePageComponent implements AfterContentInit {
     var f = this;
     f.title.setTitle("Angular Default");
     
-  }
-
-  ngAfterContentInit() {
-    var f = this;
-    
-    //
   }
 
 }